do more checks regarding directory traversals.
While we already have these regexes you can never be sure that a user
will really know what they are are doing when changing that regex.
It's better to add more check for directory traversal into the code.
Especially because our image and currDir always exists which means we
can easily use realpath to check if they are in the albumdir.

This also changes echo&&exit as well as die() to exit($message).
Additionally all $_GET vars are unset after they are validated.